+++ packages/patrol_cli/lib/src/crossplatform/app_options.dart
@@ -14,6 +14,8 @@ class FlutterAppOptions {
     required this.buildMode,
     required this.dartDefines,
     required this.dartDefineFromFilePaths,
+    required this.buildName,
+    required this.buildNumber,
   });
 
   final FlutterCommand command;
@@ -22,6 +24,8 @@ class FlutterAppOptions {
   final BuildMode buildMode;
   final Map<String, String> dartDefines;
   final List<String> dartDefineFromFilePaths;
+  final String? buildName;
+  final String? buildNumber;
 
   /// Translates these options into a proper `flutter attach`.
   @nonVirtual
@@ -225,7 +229,15 @@ class IOSAppOptions {
         if (simulator) '--simulator',
         if (noTreeShakeIcons) '--no-tree-shake-icons',
       ],
-      if (flutter.flavor != null) ...['--flavor', flutter.flavor!],
+      if (flutter.flavor case final flavor?) ...['--flavor', flavor],
+      if (flutter.buildName case final buildName?) ...[
+        '--build-name',
+        buildName,
+      ],
+      if (flutter.buildNumber case final buildNumber?) ...[
+        '--build-number',
+        buildNumber,
+      ],
       ...['--target', flutter.target],
       for (final dartDefine in flutter.dartDefines.entries) ...[
         '--dart-define',
@@ -319,7 +331,15 @@ class MacOSAppOptions {
         '--config-only',
         '--${buildMode.name}', // for example '--debug',
       ],
-      if (flutter.flavor != null) ...['--flavor', flutter.flavor!],
+      if (flutter.flavor case final flavor?) ...['--flavor', flavor],
+      if (flutter.buildName case final buildName?) ...[
+        '--build-name',
+        buildName,
+      ],
+      if (flutter.buildNumber case final buildNumber?) ...[
+        '--build-number',
+        buildNumber,
+      ],
       ...['--target', flutter.target],
       for (final dartDefine in flutter.dartDefines.entries) ...[
         '--dart-define',