+++ packages/patrol/CHANGELOG.md
@@ -1,3 +1,8 @@
+## Unreleased
+
+- Add `--build-name` and `--build-number` flags to `patrol test`, `patrol develop`, and `patrol build` commands.
+- Extend `$.native.enterText()` on Android to support `AutoCompleteTextView`. (#1162)
+
 ## 3.19.0
 
 - Fix logging for `$.native.pullToRefresh()` and `$.native.swipeBack()`. (#2707)